Separate operators from functional testing
Operator behavior should be tested independently
from the functional behavior of the components
that the operators deploy.

The primary reason for this is to preserve logs
for troubleshooting test failures and flakes in
the functional tests.
