+++ pkg/e2e-functional/tests/autoscaler/autoscaler.go
