+++ pkg/e2e-operators/tests/cluster-autoscaler-operator.go
