+++ pkg/e2e-operators/tests/utils.go
